% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.bias.circ.R
\name{diagnostic.bias.circ}
\alias{diagnostic.bias.circ}
\title{Circular bias}
\usage{
diagnostic.bias.circ(indexObs = NULL, indexPrd = NULL, obs = NULL,
  prd = NULL, dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}
}
\value{
A float number corresponding to the circular bias.
}
\description{
Function to compute the bias considering circular series
}
\author{
D. San-Martin
}
