% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.nDays.R
\name{diagnostic.nDays}
\alias{diagnostic.nDays}
\title{N days}
\usage{
diagnostic.nDays(indexObs = NULL, indexPrd = NULL, obs, prd, dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}
}
\value{
An integer corresponding to the counts
}
\description{
Function to count the sample size (number of days) used in the calculation of the indices
}
\author{
J. Bedia, D. San-Martin
}
