% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.spatial.cor.R
\name{index.spatial.cor}
\alias{index.spatial.cor}
\title{Pairwise Spatial Correlation}
\usage{
index.spatial.cor(tsl, type = "pearson", use = "pairwise.complete.obs")
}
\arguments{
\item{tsl}{A list of vectors (one for each gridpoint/station) containing the data}

\item{type}{Correlation type: \code{"pearson"} or \code{"spearman"}}

\item{use}{Check \code{?cor}. \code{"pairwise.complete.obs"} (default) ignores
uses all available info for each column pair.}
}
\value{
A matrix number corresponding to the proportion/number of days below/above the defined threshold.
}
\description{
Function to compute the pairwise spatial correlation for a set of grid
points/stations.
}
\author{
M.N. Legasa
}
