% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.spatial.laggedCrosCor.R
\name{index.spatial.laggedCrosCor}
\alias{index.spatial.laggedCrosCor}
\title{Lagged Spatiotemporal Cross-Correlations}
\usage{
index.spatial.laggedCrosCor(tsl, dates = NULL, lag = 1,
  type = "pearson", use = "pairwise.complete.obs", parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{tsl}{A list of vectors (one for each gridpoint/station) containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format
\dQuote{YYYY-MM-DD} (i.e., \code{format = "\%Y-\%m-\%d"} as in 
\code{\link{strptime}}). Note that, if dates are not provided, data 
will be considered consecutive.}

\item{lag}{Time lag. 1 by default.}

\item{type}{Correlation type: \code{"pearson"} or \code{"spearman"}}

\item{use}{Check \code{?cor}. \code{"pairwise.complete.obs"} (default) ignores
uses all available info for each column pair.}

\item{parallel}{Use parallel computation.}

\item{max.ncores}{Maximum number of cores for parallel computing.}

\item{ncores}{Number of threads for parallel computation. Default to ncores - 1.}
}
\value{
index.code matrix
}
\description{
Function to compute the pairwise lagged cross-correlations for a set of grid
points/stations.
}
\author{
M.N. Legasa

M. N. Legasa
}
