% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.rocss.R
\name{measure.rocss}
\alias{measure.rocss}
\title{Roc Skill Score (ROCSS)}
\usage{
measure.rocss(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL, dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}
}
\value{
A float number corresponding to the ROCSS
}
\description{
Calculate Roc Skill Score (ROCSS)
}
\author{
J.Baño-Medina
}
