% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation1D.R
\name{parseArgumentString}
\alias{parseArgumentString}
\title{Argument list parsing utility}
\usage{
parseArgumentString(arg.list, code, dates)
}
\arguments{
\item{arg.list}{A list of arguments in the form key=value}

\item{code}{Character string indicating the code of the target index or measure}

\item{dates}{Optional dates vector (depends of the selected index/measure)}
}
\value{
A list with two items: \code{fun}, containing the name of the validation function and \code{arg.list}, a complete list of arguments in the form key=value.
}
\description{
Internal helper to produce an argument list and a function call usable by \code{\link{do.call}} after reading metadata with
\code{\link{show.indices}} or \code{\link{show.measures}}.
}
\details{
The function is internally used by \code{\link{valueIndex1D}} and \code{\link{valueMeasure1D}}
}
\author{
J Bedia
}
\keyword{internal}
