% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_summaries.R
\name{read.lookup.table}
\alias{read.lookup.table}
\title{Read lookup tables}
\usage{
read.lookup.table(table.type)
}
\arguments{
\item{table.type}{Character string indicating what table to read. Two possible values, either
\code{"diagnostics"} or \code{"indices"}.}
}
\value{
The lookup table requested, as a \code{data.frame}.
}
\description{
Read the internal lookup tables of indices and diagnostics
}
\keyword{internal}
