% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_summaries.R
\name{show.diagnostics}
\alias{show.diagnostics}
\title{Display a summary table of VALUE diagnostics}
\usage{
show.diagnostics(diagnostic.code = NULL, function.name = NULL,
  filter = NULL)
}
\arguments{
\item{diagnostic.code}{String. Diagnostic code}

\item{function.name}{String. Name of the function}

\item{filter}{Optional character string containing a pattern to be matched}
}
\description{
Lists the available diagnostics from the VALUE Framework. A table with all diagnostics is shown on screen by default.
Diagnostics can be also queried by their VALUE codes, function name or other keywords passed to argument \code{filter}. See Details
}
\details{
The option \code{filter} allows to search by 'keywords' or 'topics'. Internally, it justs performs pattern matching on the COMMENT field 
of the corresponding lookup table. 

The original table with all the available diagnostics can be found in the following link: \url{http://www.value-cost.eu/validationportal/app/#!indices}
}
\examples{
# Get the full table of indices:
diagnostic.table <- show.diagnostics()
str(diagnostic.table)
# Search by index code:
show.diagnostics(diagnostic.code = "ts.BSFrost")
# Search by filters:
show.diagnostics(filter = "Brier")
}
\references{
Gutiérrez, J.M., Maraun, D., Widmann, M., Huth, R., Hertig, E., Benestad, R., Roessler, O., Wibig, J., Wilcke, R., Kotlarski, S., San Martín, D., Herrera, S., Bedia, J., Casanueva, A., Manzanas, R., Iturbide, M., Vrac, M., Dubrovsky, M., Ribalaygua, J., Pórtoles, J., Räty, O., Räisänen, J., Hingray, B., Raynaud, D., Casado, M.J., Ramos, P., Zerenner, T., Turco, M., Bosshard, T., Štěpánek, P., Bartholy, J., Pongracz, R., Keller, D.E., Fischer, A.M., Cardoso, R.M., Soares, P.M.M., Czernecki, B., Pagé, C., 2018. An intercomparison of a large ensemble of statistical downscaling methods over Europe: Results from the VALUE perfect predictor cross-validation experiment. International Journal of Climatology. https://doi.org/10.1002/joc.5462
}
\seealso{
show.indices
}
