% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueIndex.R
\name{valueIndex}
\alias{valueIndex}
\title{VALUE index calculation}
\usage{
valueIndex(ts, dates, indices)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). This is directly passed by the
 VALUE objects (either stations or predictions) through the element \code{object$Dates$start}.}

\item{indices}{A character vector of index codes to be computed.}
}
\value{
A named list of output indices
}
\description{
Generic funtion to compute VALUE indices from a time series
}
\examples{
obs.dataset <- file.path(find.package("VALUE"), "example_datasets", "VALUE_ECA_86_v2.zip")
obs <- loadValueStations(dataset = obs.dataset, var = "tmax", stationID = "000232")
ts <- as.vector(obs$Data)
dates <- obs$Dates$start
valueIndex(ts, dates, indices = c("AC1", "AC2", "FA25", "FA30", "FB15", "AnnualCycleAmp"))
}
