% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation1D.R
\name{valueIndex1D}
\alias{valueIndex1D}
\title{VALUE index calculation}
\usage{
valueIndex1D(ts, dates = NULL, index.codes)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). Note that the dates are not required 
by all indices, so the default is \code{NULL}. The function will yield an error if dates are required but not supplied.}

\item{index.codes}{A character vector of index codes to be computed.}
}
\value{
A named list of output indices
}
\description{
Atomic function to compute VALUE indices from a time series
}
\examples{
data(tmin.det)
valueIndex1D(tmin.det$obs, dates = tmin.det$dates,
             index.codes = c("AC1", "AC2", "FA25", "FA30", "FB15", "AnnualCycleAmp"))
}
\author{
J. Bedia
}
