% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation2D.R
\name{valueIndex2D}
\alias{valueIndex2D}
\title{VALUE spatial index calculation}
\usage{
valueIndex2D(tsl, index.code, dates = NULL, parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{tsl}{List of time series vectors.}

\item{index.code}{A character index code to be computed.}

\item{dates}{A character (or \code{POSIXct}) vector following the format
\dQuote{YYYY-MM-DD} (i.e., \code{format = "\%Y-\%m-\%d"} as in 
\code{\link{strptime}}). Note that the dates are not required 
by all indices, so the default is \code{NULL}. The function will yield an 
error if dates are required but not supplied.}

\item{parallel}{Use parallel computation.}

\item{max.ncores}{Maximum number of cores for parallel computing.}

\item{ncores}{Number of threads for parallel computation. Default to ncores - 1.}
}
\value{
index.code matrix
}
\description{
(non-atomic) Function to compute spatial VALUE indices from series.
}
\author{
M. N. Legasa
}
