% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation1D.R
\name{valueMeasure1D}
\alias{valueMeasure1D}
\title{VALUE measure calculation}
\usage{
valueMeasure1D(indexObs = NULL, indexPrd = NULL, obs = NULL,
  prd = NULL, dates = NULL, measure.codes)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}

\item{measure.codes}{A character vector of measure codes to be computed.}
}
\value{
A named list of output measures
}
\description{
Atomic function to compute VALUE measures from predicted and observed indices
}
\examples{
data(tmin.det)
# Some indices, like correlation, are computed upon the raw time series
# (i.e. no previous index is computed)
valueMeasure1D(indexObs = NULL,
               indexPrd = NULL,
               obs = tmin.det$obs,
               prd = tmin.det$pred,
               measure.codes = c("ts.rp", "ts.rs"))

# In some cases, the date information is required, with an error if omitted:

valueMeasure1D(indexObs = NULL,
               indexPrd = NULL,
               obs = tmin.det$obs,
               prd = tmin.det$pred,
               dates = tmin.det$dates,
               measure.codes = c("ts.rp", "ts.rs","ts.rpY"))

# However, most measures are computed upon previously calculated indices
indexObs <- valueIndex1D(ts = tmin.det$obs, index.codes = "AC1")
indexPrd <- valueIndex1D(ts = tmin.det$pred, index.codes = "AC1")
valueMeasure1D(indexObs = indexObs,
               indexPrd = indexPrd,
               measure.codes = c("bias", "biasRel","ratio"))
}
\author{
J. Bedia
}
